refactor: Add lazy loading to FormKlaim component in klaim/[sep]/index.vue
This commit adds lazy loading to the FormKlaim component in klaim/[sep]/index.vue. The FormKlaim component is now wrapped in a ClientOnly component, ensuring that it is only rendered when necessary. This improves the page load performance by deferring the loading of the FormKlaim component until it is actually needed.