:recycle: refactor: Update FormKlaim component to use lazy loading for tarif fields
This commit refactors the FormKlaim component in klaim/[sep]/index.vue to use lazy loading for the tarif fields. The tarif fields are now loaded dynamically, improving the page load performance by deferring the loading of the fields until they are actually needed. This optimization aims to enhance the overall user experience.