:recycle: refactor: Update Sep.vue and Ranap.vue components
Refactor Sep.vue and Ranap.vue components to improve readability and performance. This commit includes changes to the UButton component in Sep.vue to use undefined instead of an empty string for the color prop. It also updates the styling of the table headers in Ranap.vue to include whitespace-nowrap class for better display. Additionally, a new template for tgl_keluar-header is added to Ranap.vue. These changes aim to enhance the overall user experience and maintain code consistency.